refactor(ls-api): extract shared LS↔RIE API types into internal/lsapi
InvokeRequest, LogResponse, and ErrorResponse were duplicated across
cmd/localstack and cmd/ls-api. Moving them to internal/lsapi/types.go
makes both packages use the same definitions and removes the duplication
flagged in the PR review.
